app/requirements.txt
fastapi
uvicorn
python-multipart
SQLAlchemy
motor
psycopg2-binary
pymongo
cloudinary
requests
python-dotenv